Make i64

+++ ataxia_types/src/sheet/mod.rs
@@ -419,7 +419,7 @@ impl Initiative {
 #[cfg_attr(feature = "serde", serde(default))]
 #[cfg_attr(feature = "rhai", derive(CustomType))]
 /// Walking speed in feet.
-pub struct WalkingSpeed(pub u32);
+pub struct WalkingSpeed(pub i64);
 
 impl Default for WalkingSpeed {
     fn default() -> Self {
